requirements.txt
apache-airflow
kafka-python
